Adiciona desconto para pagamento em dinheiro
